refactor : ohho storage module 생성
