+++ ohho-storage/src/main/java/kr/ohho/config/JpaConfig.java
@@ -1,0 +1,10 @@
+package kr.ohho.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@EnableJpaAuditing
+@Configuration
+public class JpaConfig {
+
+}
