+++ ohho-storage/src/main/java/kr/ohho/domain/partner/repository/PartnerCustomRepository.java
@@ -1,0 +1,8 @@
+package kr.ohho.domain.partner.repository;
+
+import java.time.LocalDate;
+
+public interface PartnerCustomRepository {
+
+    Long updateDdayAndGoal(Long id, LocalDate dDay, String goal);
+}
