+++ ohho-storage/src/main/java/kr/ohho/domain/partner/repository/PartnerRepository.java
@@ -1,0 +1,8 @@
+package kr.ohho.domain.partner.repository;
+
+import kr.ohho.domain.partner.Partner;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PartnerRepository extends JpaRepository<Partner, Long> {
+
+}
