+++ ohho-storage/src/main/java/kr/ohho/seed/R__001_SEED_exercise.sql
