+++ ohho-storage/src/test/java/kr/ohho/TestConfiguration.java
@@ -1,0 +1,8 @@
+package kr.ohho;
+
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TestConfiguration {
+
+}
