--- src/main/java/ohho/backend/spring/common/entities/BaseEntity.java
@@ -1,35 +1,0 @@
-package ohho.backend.spring.common.entities;
-
-import java.time.LocalDateTime;
-import javax.persistence.Column;
-import javax.persistence.EntityListeners;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.MappedSuperclass;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.ToString;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-@Getter
-@MappedSuperclass
-@EntityListeners(AuditingEntityListener.class)
-@ToString(of = {"id"})
-@EqualsAndHashCode(of = "id")
-public class BaseEntity {
-
-    @Id
-    @GeneratedValue
-    private Long id;
-
-    /* date 컬럼 */
-
-    @CreatedDate
-    @Column(updatable = false, nullable = false)
-    private LocalDateTime createdAt;
-
-    @LastModifiedDate
-    private LocalDateTime updatedAt;
-}
