--- src/main/java/ohho/backend/spring/domain/member/repository/MemberRepository.java
@@ -1,21 +1,0 @@
-package ohho.backend.spring.domain.member.repository;
-
-import java.util.List;
-import java.util.Optional;
-import ohho.backend.spring.domain.member.entities.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MemberRepository extends JpaRepository<Member, Long> {
-
-    Optional<Member> findByEmail(String email);
-
-    Optional<Member> findByEmailAndPassword(String email, String password);
-
-    Optional<Member> findByNickname(String nickname);
-
-    Optional<Member> findByGoogleUserId(String googleUserId);
-
-    boolean existsByEmail(String email);
-
-    boolean existsByNickname(String nickname);
-}
