--- src/main/java/ohho/backend/spring/domain/partner/repository/PartnerCustomRepository.java
@@ -1,8 +1,0 @@
-package ohho.backend.spring.domain.partner.repository;
-
-import java.time.LocalDate;
-
-public interface PartnerCustomRepository {
-
-    Long updateDdayAndGoal(Long id, LocalDate dDay, String goal);
-}
